Created an evidence scheme and added TurbiniaEvidenceCollection to redis
